refactor: rename to exchange-price-exporter
