+++ assets/prometheus.yml
@@ -4,4 +4,4 @@ global:
 scrape_configs:
   - job_name: 'prices'
     static_configs:
-      - targets: ['terra-price-exporter:8000']
+      - targets: ['exchange-price-exporter:8000']
