--- terra_price_exporter/exchange/__init__.py
@@ -1,4 +1,0 @@
-from terra_price_exporter.exchange.bithumb import Bithumb
-from terra_price_exporter.exchange.coinone import Coinone
-
-exchanges = dict(bithumb=Bithumb, coinone=Coinone)
