+++ tests/test_exchange_price_exporter.py
@@ -1,4 +1,4 @@
-from terra_price_exporter import __version__
+from exchange_price_exporter import __version__
 
 
 def test_version():
